Fix particle color gradients when using texture atlas particles

+++ examples/local_space.rs
@@ -108,7 +108,7 @@ pub fn circler(
     time: Res<Time>,
     mut particle_system_query: Query<(&Circler, &mut Transform), With<ParticleSystem>>,
 ) {
-    let rad = time.elapsed_seconds() as f32;
+    let rad = time.elapsed_seconds();
     let quat = Quat::from_rotation_z(rad).normalize();
     let dir = quat * Vec3::Y;
     for (circler, mut transform) in &mut particle_system_query {
